Applied second round of Keyvan feedback
