+++ src/app/course/_test/mock.ts
@@ -12,6 +12,53 @@ import {
 import {TokenUse} from "@app/_models/token_use"
 import {MOCK_CATEGORY, MOCK_COURSE_EVENT} from "@app/problems/_test/mock"
 import {Stats} from "@app/_models/user_difficulty_stats"
+import {Goal, GoalItem} from "@app/_models/goal/goal"
+
+export const MOCK_GOAL_ITEM: GoalItem = {
+    category: 71,
+    difficulty: "EASY",
+    number_of_questions: 5
+}
+
+export const MOCK_GOAL_ITEM_2: GoalItem = {
+    category: 72,
+    difficulty: "MEDIUM",
+    number_of_questions: 15
+}
+
+export const MOCK_GOAL_ITEM_3: GoalItem = {
+    category: 71,
+    difficulty: "HARD",
+    number_of_questions: 20
+}
+
+export const MOCK_GOAL_ITEMS: GoalItem[] = [MOCK_GOAL_ITEM, MOCK_GOAL_ITEM_2, MOCK_GOAL_ITEM_3]
+
+export const MOCK_GOAL: Goal = {
+    id: 13,
+    course_reg: 1,
+    start_date: "2004-09-03T20:50:00-07:00",
+    end_date: "2022-11-10T16:40:37.860661-08:00",
+    progress: 0,
+    is_finished: true,
+    number_of_questions: 40,
+    goal_items: MOCK_GOAL_ITEMS,
+    stats: null,
+}
+
+export const MOCK_GOAL_2: Goal = {
+    id: 14,
+    course_reg: 1,
+    start_date: "2006-03-04T20:50:00-07:00",
+    end_date: "2021-21-13T16:40:37.860661-08:00",
+    progress: 0,
+    is_finished: false,
+    number_of_questions: 40,
+    goal_items: MOCK_GOAL_ITEMS,
+    stats: null,
+}
+
+export const MOCK_GOALS: Goal[] = [MOCK_GOAL, MOCK_GOAL_2]
 
 export const MOCK_TOKEN_USE_OPTION1: TokenUseOption = {
     id: 1,
