+++ src/app/course/goal/goal-create/goal-creation-page.component.spec.ts
@@ -1,12 +1,13 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing'
-import {GoalCreateComponent} from "@app/course/goal-create/goal-creation-page.component"
-import {MOCK_GOALS, MOCK_CATEGORIES, MOCK_DIFFICULTIES} from "@app/problems/_test/mock"
+import {GoalCreateComponent} from "@app/course/goal/goal-create/goal-creation-page.component"
+import {MOCK_GOALS} from "@app/course/_test/mock"
+import {MOCK_CATEGORIES, MOCK_DIFFICULTIES} from "@app/problems/_test/mock"
 import {CategoryService} from "@app/_services/api/category.service"
 import {CategoryServiceMock} from "@test/category.service.mock"
 import {DifficultyService} from "@app/problems/_services/difficulty.service"
 import {DifficultyServiceMock} from "@app/problems/_test/_services/difficulty.service.mock"
 import {GoalService} from "@app/course/_services/goal.service"
-import {GoalServiceMock} from "@test/goal.service.mock"
+import {GoalServiceMock} from "@app/course/_test/_services/goal.service.mock"
 import {ReactiveFormsModule} from "@angular/forms"
 import {RouterModule} from "@angular/router"
 import {
